Move common build system files to build-system/common/
