+++ build-system/common/check-package-manager.js
