+++ build-system/common/ctrlcHandler.js
