+++ build-system/tasks/babel-plugin-tests.js
@@ -17,7 +17,7 @@
 
 const gulp = require('gulp');
 const jest = require('gulp-jest').default;
-const {isTravisBuild} = require('../travis');
+const {isTravisBuild} = require('../common/travis');
 
 /**
  * Simple wrapper around the jest tests for custom babel plugins.
