+++ build-system/tasks/bundle-size.js
@@ -25,15 +25,19 @@ const path = require('path');
 const requestPost = BBPromise.promisify(require('request').post);
 const url = require('url');
 const {
+  gitCommitHash,
+  gitTravisMasterBaseline,
+  shortSha,
+} = require('../common/git');
+const {
   isTravisPullRequestBuild,
   isTravisPushBuild,
   travisPushBranch,
   travisRepoSlug,
-} = require('../travis');
+} = require('../common/travis');
 const {
   VERSION: internalRuntimeVersion,
 } = require('../compile/internal-version');
-const {gitCommitHash, gitTravisMasterBaseline, shortSha} = require('../git');
 
 const runtimeFile = './dist/v0.js';
 const normalizedRtvNumber = '1234567890123';
