+++ build-system/tasks/dev-dashboard-tests.js
@@ -18,7 +18,7 @@
 const config = require('../test-configs/config');
 const deglob = require('globs-to-files');
 const Mocha = require('mocha');
-const {isTravisBuild} = require('../travis');
+const {isTravisBuild} = require('../common/travis');
 
 /**
  * Run all the dev dashboard tests
