+++ build-system/tasks/presubmit-checks.js
@@ -109,9 +109,8 @@ const forbiddenTerms = {
       'If you run against this, use console/*OK*/.[log|error] to ' +
       'whitelist a legit case.',
     whitelist: [
-      'build-system/server/app.js',
-      'build-system/server/amp4test.js',
-      'build-system/check-package-manager.js',
+      'build-system/common/check-package-manager.js',
+      'build-system/compile/single-pass.js',
       'build-system/pr-check/build.js',
       'build-system/pr-check/build-targets.js',
       'build-system/pr-check/checks.js',
@@ -125,13 +124,14 @@ const forbiddenTerms = {
       'build-system/pr-check/validator-tests.js',
       'build-system/pr-check/visual-diff-tests.js',
       'build-system/pr-check/yarn-checks.js',
+      'build-system/server/app.js',
+      'build-system/server/amp4test.js',
       'build-system/tasks/build.js',
       'build-system/tasks/check-exact-versions.js',
       'build-system/tasks/check-types.js',
       'build-system/tasks/dist.js',
       'build-system/tasks/generate-runner.js',
       'build-system/tasks/helpers.js',
-      'build-system/compile/single-pass.js',
       'validator/nodejs/index.js', // NodeJs only.
       'validator/engine/parse-css.js',
       'validator/engine/validator-in-browser.js',
@@ -1161,10 +1161,10 @@ const forbiddenTermsSrcInclusive = {
   "process\\.env(\\.TRAVIS|\\[\\'TRAVIS)": {
     message:
       'Do not directly use process.env.TRAVIS. Instead, add a ' +
-      'function to build-system/travis.js',
+      'function to build-system/common/travis.js',
     whitelist: [
-      'build-system/check-package-manager.js',
-      'build-system/travis.js',
+      'build-system/common/check-package-manager.js',
+      'build-system/common/travis.js',
     ],
   },
   '\\.matches\\(': 'Please use matches() helper in src/dom.js',
