+++ package.json
@@ -15,7 +15,7 @@
     "url": "https://github.com/ampproject/amphtml.git"
   },
   "scripts": {
-    "preinstall": "node build-system/check-package-manager.js"
+    "preinstall": "node build-system/common/check-package-manager.js"
   },
   "dependencies": {
     "@ampproject/animations": "0.2.0",
